requirements.txt
Flask==3.0.0
MongoEngine==0.28.2
python-dotenv==1.0.0
requests==2.32.3
PyMuPDF==1.24.5
pdfminer.six==20231228
spacy>=3.8.0
transformers==4.40.0
pymongo==4.7.2
gunicorn==22.0.0
bootstrap-flask==2.3.3
chartjs==1.2
google-generativeai>=0.3.0
werkzeug>=2.0.0
markdown2==2.4.10
reportlab==4.0.4